requirements.txt
paho-mqtt==1.6.1
aiocoap==0.4.14
Flask==3.1.0
Flask-RESTful==0.3.10
python-dotenv==1.1.0